Remove index for pages, to simplify imports and remove bloat
