+++ src/App.js
@@ -6,8 +6,8 @@ import store from "utils/store"
 import "utils/ignore"
 
 // assets
-import { imageAssets } from "theme/images"
-import { fontAssets } from "theme/fonts"
+import { imageAssets } from "./theme/images"
+import { fontAssets } from "./theme/fonts"
 import Navigator from "./navigator"
 
 const App = () => {
