+++ src/navigator/Drawer/DrawerMenu.js
@@ -4,7 +4,7 @@ import { View, SafeAreaView, Text } from "react-native"
 
 import { DrawerActions } from "@react-navigation/native"
 import FontIcon from "react-native-vector-icons/FontAwesome5"
-import { colors } from "theme"
+import { colors } from "../../theme"
 
 const styles = {
   root: {
