+++ src/navigator/Stacks/Stacks.js
@@ -1,17 +1,17 @@
 import React from "react"
 import { createStackNavigator } from "@react-navigation/stack"
-import { colors } from "theme"
-import Details from "pages/Details"
-import Order from "pages/Order"
-import Me from "pages/Me"
 import HeaderLeft from "./HeaderLeft"
 import HeaderTitle from "./HeaderTitle"
+import { colors } from "../../theme"
+import Details from "../../pages/Details"
+import Order from "../../pages/Order"
+import Me from "../../pages/UserProfile"
 import Logo from "../../pages/Logo"
 import Login from "../../pages/Login"
 import PFPSelector from "../../pages/PFPSelector"
-import Fun from "../../pages/Fun"
+import Fun from "../../pages/Discover"
 import Checkout from "../../pages/Checkout"
-import DietaryRestrictions from "../../pages/Dietary Restrictions"
+import DietaryRestrictions from "../../pages/DietaryRestrictions"
 import Home from "../../pages/Home"
 import CreateAccount from "../../pages/CreateAccount"
 import Restaurant from "../../pages/Restaurant"
