--- src/pages/Checkout/index.js
@@ -1,3 +1,0 @@
-import Checkout from "./Checkout"
-
-export default Checkout
