--- src/pages/CreateAccount/index.js
@@ -1,3 +1,0 @@
-import CreateAccount from "./CreateAccount"
-
-export default CreateAccount
