--- src/pages/Details/index.js
@@ -1,3 +1,0 @@
-import Details from "./Details"
-
-export default Details
