--- src/pages/Dietary Restrictions/index.js
@@ -1,3 +1,0 @@
-import DietaryRestrictions from "./DietaryRestrictions"
-
-export default DietaryRestrictions
