--- src/pages/Fun/index.js
@@ -1,3 +1,0 @@
-import Fun from "./Fun"
-
-export default Fun
