+++ src/pages/Home.tsx
@@ -1,12 +1,10 @@
 import React from "react"
 import PropTypes from "prop-types"
-import {
-  StyleSheet, StatusBar, SafeAreaView, Image,
-} from "react-native"
+import { StyleSheet, StatusBar, SafeAreaView, Image } from "react-native"
 import RestaurantCard from "components/Restaurant Card"
 import { ScrollView } from "react-native-gesture-handler"
 import { vw } from "react-native-expo-viewport-units"
-import { images } from "../../theme"
+import { images } from "../theme"
 
 const styles = StyleSheet.create({
   root: {
