--- src/pages/Logo/index.js
@@ -1,3 +1,0 @@
-import Logo from "./Logo"
-
-export default Logo
