--- src/pages/Me/index.js
@@ -1,3 +1,0 @@
-import Me from "./Me"
-
-export default Me
