--- src/pages/Order/index.js
@@ -1,3 +1,0 @@
-import Order from "./Order"
-
-export default Order
