+++ src/pages/PFPSelector.tsx
@@ -1,9 +1,9 @@
 import React from "react"
 import PropTypes from "prop-types"
 import { StyleSheet, Text, View } from "react-native"
-import { colors } from "theme"
-import PFP from "../../components/PFP"
-import Button from "../../components/Button"
+import { colors } from "../theme"
+import PFP from "../components/PFP"
+import Button from "../components/Button"
 
 const styles = StyleSheet.create({
   root: {
