--- src/pages/PFPSelector/index.js
@@ -1,3 +1,0 @@
-import PFPSelector from "./PFPSelector"
-
-export default PFPSelector
