+++ src/pages/Restaurant.tsx
@@ -1,13 +1,11 @@
 import React from "react"
 import PropTypes from "prop-types"
-import {
-  Text, View, StatusBar, Image, SafeAreaView,
-} from "react-native"
+import { Text, View, StatusBar, Image, SafeAreaView } from "react-native"
 import { ScrollView } from "react-native-gesture-handler"
 import EStyleSheet from "react-native-extended-stylesheet"
 import { vw } from "react-native-expo-viewport-units"
-import RestaurantItem from "../../components/RestaurantItem"
-import { images } from "../../theme"
+import RestaurantItem from "../components/RestaurantItem"
+import { images } from "../theme"
 
 const styles = EStyleSheet.create({
   topImage: {
