--- src/pages/Restaurant/index.js
@@ -1,3 +1,0 @@
-import Restaurant from "./Restaurant"
-
-export default Restaurant
